ai-model/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
opencv-python==4.9.0.80
numpy==1.24.3
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pgvector==0.2.4
Pillow==10.2.0
# Using specific TensorFlow version for compatibility
tensorflow==2.15.0
# Using older DeepFace version that works with TensorFlow 2.15
deepface==0.0.75
# Protobuf version compatible with TensorFlow 2.15
protobuf==3.20.3
